Link goalsServices with the HomeView. Choose the color of units based on the goals.

+++ units-tracking-app/units-tracking-app/BusinessLayer/ GoalsService.swift
@@ -10,12 +10,12 @@ import Foundation
 
 class  GoalsService: ObservableObject {
     /// Returns the number of units per day which is the max value but still within a limit.
-    func getUnitsPerDay() -> Int {
-        return 3
+    func getUnitsPerDay() -> Double {
+        return 3.0
     }
     
     /// Returns the number of units per 7 days which is the max value but still within a limit.
-    func getUnitsPer7Days() -> Int {
-        return 14
+    func getUnitsPer7Days() -> Double {
+        return 14.0
     }
 }
